Fix WGC DXGI access and WinUI layout for VS 2026 /permissive-.
Inbox interop headers omit IDirect3DDxgiInterfaceAccess; std::max(int, LONG) does not deduce.

+++ Apps/WindowsStudio/Sources/WindowsCapture/winui_shell.cpp
@@ -560,8 +560,8 @@ void layoutChrome(HWND hwnd) {
     }
     RECT rc{};
     GetClientRect(hwnd, &rc);
-    const int width = (std::max)(1, rc.right - rc.left);
-    const int height = (std::max)(1, rc.bottom - rc.top);
+    const int width = (std::max)(1, static_cast<int>(rc.right - rc.left));
+    const int height = (std::max)(1, static_cast<int>(rc.bottom - rc.top));
     const int islandH = 220;
     if (gUi.island) {
         SetWindowPos(gUi.island, HWND_TOP, 0, 0, width, islandH, SWP_SHOWWINDOW);
